test: get generated cms_test.exs working
involves handling the author association on pages.
I added a support file with a user_author_fixture
function that first creates a user in order to
hang an author off it.
I added default:0 to views attribute of page schema,
because by default a Repo.insert returns the structs
passed to the insert, not the actual values in db
after insert.
On CMS.create_page, when the result is {:ok, page},
I do Repo.preload(author: [user: :credential]) on it.