test: fix cms/page_controller_test to handle authentication
in setup, use Support.CMS.Fixtures.user_fixture to make a user, and
set its ide in the session of the conn.

set the authenticated user in the test context as well.
In setup(:create_page), get the authenticated_user and
pass it to Fixtures.user_author_fixture to get the author
of the authenticated user, and use it to create page.